[CDF-21838] S🪓 plit utils.py into a submodule (#1167)
* refactor; split utils.py into a submodule

* tests: updated tests
