+++ cognite_toolkit/_cdf_tk/utils/auth.py
@@ -13,28 +13,17 @@
 # limitations under the License.
 from __future__ import annotations
 
-import difflib
-import hashlib
 import itertools
 import json
-import logging
 import os
-import re
 import shutil
-import sys
-import tempfile
-import typing
-from abc import abstractmethod
-from collections import UserDict, defaultdict
-from collections.abc import Collection, ItemsView, Iterable, Iterator, KeysView, MutableSequence, Sequence, ValuesView
-from contextlib import contextmanager
+from collections.abc import Sequence
 from dataclasses import dataclass, field, fields
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, Literal, Optional, TypeVar, overload
+from typing import TYPE_CHECKING, Any, Literal, TypeAlias, overload
 
 import questionary
 import typer
-import yaml
 from cognite.client import ClientConfig
 from cognite.client.config import global_config
 from cognite.client.credentials import (
@@ -44,7 +33,6 @@ from cognite.client.credentials import (
     OAuthInteractive,
     Token,
 )
-from cognite.client.data_classes import CreatedSession
 from cognite.client.data_classes.capabilities import (
     AssetsAcl,
     Capability,
@@ -54,7 +42,6 @@ from cognite.client.data_classes.capabilities import (
     SecurityCategoriesAcl,
     TimeSeriesAcl,
 )
-from cognite.client.data_classes.data_modeling import DataModelId, View, ViewId
 from cognite.client.data_classes.iam import TokenInspection
 from cognite.client.exceptions import CogniteAPIError
 from questionary import Choice
@@ -62,11 +49,8 @@ from rich import print
 from rich.prompt import Prompt
 
 from cognite_toolkit._cdf_tk.client import ToolkitClient
-from cognite_toolkit._cdf_tk.client.testing import ToolkitClientMock
 from cognite_toolkit._cdf_tk.constants import (
     _RUNNING_IN_BROWSER,
-    BUILTIN_MODULES,
-    ROOT_MODULES,
     TOOLKIT_CLIENT_ENTRA_ID,
     URL,
 )
@@ -74,25 +58,13 @@ from cognite_toolkit._cdf_tk.exceptions import (
     AuthenticationError,
     AuthorizationError,
     ResourceRetrievalError,
-    ToolkitError,
     ToolkitResourceMissingError,
-    ToolkitYAMLFormatError,
 )
 from cognite_toolkit._cdf_tk.tk_warnings import MediumSeverityWarning
 from cognite_toolkit._version import __version__
 
-if sys.version_info < (3, 10):
-    from typing_extensions import TypeAlias
-else:
-    from typing import TypeAlias
-
-
 if TYPE_CHECKING:
-    from sentry_sdk.types import Event as SentryEvent
-    from sentry_sdk.types import Hint as SentryHint
-
-
-logger = logging.getLogger(__name__)
+    pass
 
 
 LoginFlow: TypeAlias = Literal["client_credentials", "token", "device_code", "interactive"]
@@ -1082,775 +1054,3 @@ class CDFToolConfig:
             return self._cache.locationfilter_id_by_external_id[external_id]
         else:
             return [self._cache.locationfilter_id_by_external_id[ext_id] for ext_id in external_id]
-
-
-@overload
-def load_yaml_inject_variables(
-    filepath: Path | str, variables: dict[str, str | None], required_return_type: Literal["list"], validate: bool = True
-) -> list[dict[str, Any]]: ...
-
-
-@overload
-def load_yaml_inject_variables(
-    filepath: Path | str, variables: dict[str, str | None], required_return_type: Literal["dict"], validate: bool = True
-) -> dict[str, Any]: ...
-
-
-@overload
-def load_yaml_inject_variables(
-    filepath: Path | str,
-    variables: dict[str, str | None],
-    required_return_type: Literal["any"] = "any",
-    validate: bool = True,
-) -> dict[str, Any] | list[dict[str, Any]]: ...
-
-
-def load_yaml_inject_variables(
-    filepath: Path | str,
-    variables: dict[str, str | None],
-    required_return_type: Literal["any", "list", "dict"] = "any",
-    validate: bool = True,
-) -> dict[str, Any] | list[dict[str, Any]]:
-    if isinstance(filepath, str):
-        content = filepath
-    else:
-        content = filepath.read_text()
-    for key, value in variables.items():
-        if value is None:
-            continue
-        content = content.replace(f"${{{key}}}", value)
-    if validate:
-        for match in re.finditer(r"\$\{([^}]+)\}", content):
-            environment_variable = match.group(1)
-            suffix = f" It is expected in {filepath.name}." if isinstance(filepath, Path) else ""
-            MediumSeverityWarning(
-                f"Variable {environment_variable} is not set in the environment.{suffix}"
-            ).print_warning()
-
-    if yaml.__with_libyaml__:
-        # CSafeLoader is faster than yaml.safe_load
-        result = yaml.CSafeLoader(content).get_data()
-    else:
-        result = yaml.safe_load(content)
-    if required_return_type == "any":
-        return result
-    elif required_return_type == "list":
-        if isinstance(result, list):
-            return result
-        raise ValueError(f"Expected a list, but got {type(result)}")
-    elif required_return_type == "dict":
-        if isinstance(result, dict):
-            return result
-        raise ValueError(f"Expected a dict, but got {type(result)}")
-    else:
-        raise ValueError(f"Unknown required_return_type {required_return_type}")
-
-
-@overload
-def read_yaml_file(filepath: Path, expected_output: Literal["dict"] = "dict") -> dict[str, Any]: ...
-
-
-@overload
-def read_yaml_file(filepath: Path, expected_output: Literal["list"]) -> list[dict[str, Any]]: ...
-
-
-def read_yaml_file(
-    filepath: Path, expected_output: Literal["list", "dict"] = "dict"
-) -> dict[str, Any] | list[dict[str, Any]]:
-    """Read a YAML file and return a dictionary
-
-    filepath: path to the YAML file
-    """
-    try:
-        config_data = read_yaml_content(filepath.read_text())
-    except yaml.YAMLError as e:
-        print(f"  [bold red]ERROR:[/] reading {filepath}: {e}")
-        return {}
-
-    if expected_output == "list" and isinstance(config_data, dict):
-        ToolkitYAMLFormatError(f"{filepath} did not contain `list` as expected")
-    elif expected_output == "dict" and isinstance(config_data, list):
-        ToolkitYAMLFormatError(f"{filepath} did not contain `dict` as expected")
-    return config_data
-
-
-def read_yaml_content(content: str) -> dict[str, Any] | list[dict[str, Any]]:
-    """Read a YAML string and return a dictionary
-
-    content: string containing the YAML content
-    """
-    if yaml.__with_libyaml__:
-        # CSafeLoader is faster than yaml.safe_load
-        config_data = yaml.CSafeLoader(content).get_data()
-    else:
-        config_data = yaml.safe_load(content)
-    return config_data
-
-
-def calculate_directory_hash(
-    directory: Path,
-    exclude_prefixes: set[str] | None = None,
-    ignore_files: set[str] | None = None,
-    shorten: bool = False,
-) -> str:
-    sha256_hash = hashlib.sha256()
-
-    # Walk through each file in the directory
-    for filepath in sorted(directory.rglob("*"), key=lambda p: str(p.relative_to(directory))):
-        if filepath.is_dir():
-            continue
-        if exclude_prefixes and any(filepath.name.startswith(prefix) for prefix in exclude_prefixes):
-            continue
-        if ignore_files and filepath.suffix in ignore_files:
-            continue
-        relative_path = filepath.relative_to(directory)
-        sha256_hash.update(relative_path.as_posix().encode("utf-8"))
-        # Open each file and update the hash
-        with filepath.open("rb") as file:
-            while chunk := file.read(8192):
-                # Get rid of Windows line endings to make the hash consistent across platforms.
-                sha256_hash.update(chunk.replace(b"\r\n", b"\n"))
-
-    calculated = sha256_hash.hexdigest()
-    if shorten:
-        return calculated[:8]
-    return calculated
-
-
-def calculate_secure_hash(item: dict[str, Any]) -> str:
-    """Calculate a secure hash of a dictionary"""
-    sha256_hash = hashlib.sha512(usedforsecurity=True)
-    sha256_hash.update(json.dumps(item, sort_keys=True).encode("utf-8"))
-    return sha256_hash.hexdigest()
-
-
-def calculate_str_or_file_hash(content: str | Path, shorten: bool = False) -> str:
-    sha256_hash = hashlib.sha256()
-    if isinstance(content, Path):
-        content = safe_read(content)
-    # Get rid of Windows line endings to make the hash consistent across platforms.
-    sha256_hash.update(content.encode("utf-8").replace(b"\r\n", b"\n"))
-    calculated = sha256_hash.hexdigest()
-    if shorten:
-        return calculated[:8]
-    return calculated
-
-
-def calculate_bytes_or_file_hash(content: bytes | Path, shorten: bool = False) -> str:
-    sha256_hash = hashlib.sha256()
-    if isinstance(content, Path):
-        # Get rid of Windows line endings to make the hash consistent across platforms.
-        content = content.read_bytes().replace(b"\r\n", b"\n")
-    sha256_hash.update(content)
-    calculated = sha256_hash.hexdigest()
-    if shorten:
-        return calculated[:8]
-    return calculated
-
-
-def get_oneshot_session(client: ToolkitClient) -> CreatedSession | None:
-    """Get a oneshot (use once) session for execution in CDF"""
-    # Special case as this utility function may be called with a new client created in code,
-    # it's hard to mock it in tests.
-    if isinstance(client, ToolkitClientMock):
-        bearer = "123"
-    else:
-        (_, bearer) = client.config.credentials.authorization_header()
-    ret = client.post(
-        url=f"/api/v1/projects/{client.config.project}/sessions",
-        json={
-            "items": [
-                {
-                    "oneshotTokenExchange": True,
-                },
-            ],
-        },
-        headers={"Authorization": bearer},
-    )
-    if ret.status_code == 200:
-        return CreatedSession.load(ret.json()["items"][0])
-    return None
-
-
-@dataclass(frozen=True)
-class YAMLComment:
-    """This represents a comment in a YAML file. It can be either above or after a variable."""
-
-    above: list[str] = field(default_factory=list)
-    after: list[str] = field(default_factory=list)
-
-    @property
-    def comment(self) -> str:
-        return "\n".join(self.above) + "\n" + "\n".join(self.after)
-
-
-T_Key = TypeVar("T_Key")
-T_Value = TypeVar("T_Value")
-
-
-class YAMLWithComments(UserDict[T_Key, T_Value]):
-    @staticmethod
-    def _extract_comments(raw_file: str, key_prefix: tuple[str, ...] = tuple()) -> dict[tuple[str, ...], YAMLComment]:
-        """Extract comments from a raw file and return a dictionary with the comments."""
-        comments: dict[tuple[str, ...], YAMLComment] = defaultdict(YAMLComment)
-        position: Literal["above", "after"]
-        init_value: object = object()
-        variable: str | None | object = init_value
-        last_comments: list[str] = []
-        last_variable: str | None = None
-        last_leading_spaces = 0
-        parent_variables: list[str] = []
-        indent: int | None = None
-        for line in raw_file.splitlines():
-            if ":" in line:
-                # Is variable definition
-                leading_spaces = len(line) - len(line.lstrip())
-                variable = str(line.split(":", maxsplit=1)[0].strip())
-                if leading_spaces > last_leading_spaces and last_variable:
-                    parent_variables.append(last_variable)
-                    if indent is None:
-                        # Automatically indent based on the first variable
-                        indent = leading_spaces
-                elif leading_spaces < last_leading_spaces and parent_variables:
-                    parent_variables = parent_variables[: -((last_leading_spaces - leading_spaces) // (indent or 2))]
-
-                if last_comments:
-                    comments[(*key_prefix, *parent_variables, variable)].above.extend(last_comments)
-                    last_comments.clear()
-
-                last_variable = variable
-                last_leading_spaces = leading_spaces
-
-            if "#" in line:
-                # Potentially has comment.
-                before, comment = str(line).rsplit("#", maxsplit=1)
-                position = "after" if ":" in before else "above"
-                if position == "after" and (before.count('"') % 2 == 1 or before.count("'") % 2 == 1):
-                    # The comment is inside a string
-                    continue
-                # This is a new comment.
-                if (position == "after" or variable is None) and variable is not init_value:
-                    key = (*key_prefix, *parent_variables, *((variable and [variable]) or []))  # type: ignore[misc]
-                    if position == "after":
-                        comments[key].after.append(comment.strip())
-                    else:
-                        comments[key].above.append(comment.strip())
-                else:
-                    last_comments.append(comment.strip())
-
-        return dict(comments)
-
-    def _dump_yaml_with_comments(self, indent_size: int = 2, newline_after_indent_reduction: bool = False) -> str:
-        """Dump a config dictionary to a yaml string"""
-        config = self.dump()
-        dumped = yaml.dump(config, sort_keys=False, indent=indent_size)
-        out_lines = []
-        if comments := self._get_comment(tuple()):
-            for comment in comments.above:
-                out_lines.append(f"# {comment}")
-        last_indent = 0
-        last_variable: str | None = None
-        path: tuple[str, ...] = tuple()
-        for line in dumped.splitlines():
-            indent = len(line) - len(line.lstrip())
-            if last_indent < indent:
-                if last_variable is None:
-                    raise ValueError("Unexpected state of last_variable being None")
-                path = (*path, last_variable)
-            elif last_indent > indent:
-                if newline_after_indent_reduction:
-                    # Adding some extra space between modules
-                    out_lines.append("")
-                indent_reduction_steps = (last_indent - indent) // indent_size
-                path = path[:-indent_reduction_steps]
-
-            variable = line.split(":", maxsplit=1)[0].strip()
-            if comments := self._get_comment((*path, variable)):
-                for line_comment in comments.above:
-                    out_lines.append(f"{' ' * indent}# {line_comment}")
-                if after := comments.after:
-                    line = f"{line} # {after[0]}"
-
-            out_lines.append(line)
-            last_indent = indent
-            last_variable = variable
-        out_lines.append("")
-        return "\n".join(out_lines)
-
-    @abstractmethod
-    def dump(self) -> dict[str, Any]: ...
-
-    @abstractmethod
-    def _get_comment(self, key: tuple[str, ...]) -> YAMLComment | None: ...
-
-    # This is to get better type hints in the IDE
-    def items(self) -> ItemsView[T_Key, T_Value]:
-        return super().items()
-
-    def keys(self) -> KeysView[T_Key]:
-        return super().keys()
-
-    def values(self) -> ValuesView[T_Value]:
-        return super().values()
-
-
-def retrieve_view_ancestors(client: ToolkitClient, parents: list[ViewId], cache: dict[ViewId, View]) -> list[View]:
-    """Retrieves all ancestors of a view.
-
-    This will mutate the cache that is passed in, and return a list of views that are the ancestors of the views in the parents list.
-
-    Args:
-        client: The Cognite client to use for the requests
-        parents: The parents of the view to retrieve all ancestors for
-        cache: The cache to store the views in
-    """
-    parent_ids = parents
-    found: list[View] = []
-    while parent_ids:
-        to_lookup = []
-        grand_parent_ids = []
-        for parent in parent_ids:
-            if parent in cache:
-                found.append(cache[parent])
-                grand_parent_ids.extend(cache[parent].implements or [])
-            else:
-                to_lookup.append(parent)
-
-        if to_lookup:
-            looked_up = client.data_modeling.views.retrieve(to_lookup)
-            cache.update({view.as_id(): view for view in looked_up})
-            found.extend(looked_up)
-            for view in looked_up:
-                grand_parent_ids.extend(view.implements or [])
-
-        parent_ids = grand_parent_ids
-    return found
-
-
-def sentry_exception_filter(event: SentryEvent, hint: SentryHint) -> Optional[SentryEvent]:
-    if "exc_info" in hint:
-        exc_type, exc_value, tb = hint["exc_info"]
-        # Returning None prevents the event from being sent to Sentry
-        if isinstance(exc_value, ToolkitError):
-            return None
-    return event
-
-
-@contextmanager
-def tmp_build_directory() -> typing.Generator[Path, None, None]:
-    build_dir = Path(tempfile.mkdtemp(prefix="build.", suffix=".tmp", dir=Path.cwd()))
-    try:
-        yield build_dir
-    finally:
-        shutil.rmtree(build_dir)
-
-
-def flatten_dict(dct: dict[str, Any]) -> dict[tuple[str, ...], Any]:
-    """Flatten a dictionary to a list of tuples with the key path and value."""
-    items: dict[tuple[str, ...], Any] = {}
-    for key, value in dct.items():
-        if isinstance(value, dict):
-            for sub_key, sub_value in flatten_dict(value).items():
-                items[(key, *sub_key)] = sub_value
-        else:
-            items[(key,)] = value
-    return items
-
-
-def iterate_modules(root_dir: Path) -> Iterator[tuple[Path, list[Path]]]:
-    """Iterate over all modules in the project and yield the module directory and all files in the module.
-
-    Args:
-        root_dir (Path): The root directory of the project
-
-    Yields:
-        Iterator[tuple[Path, list[Path]]]: A tuple containing the module directory and a list of all files in the module
-
-    """
-    if root_dir.name in ROOT_MODULES:
-        yield from _iterate_modules(root_dir)
-        return
-    elif root_dir.name == BUILTIN_MODULES:
-        yield from _iterate_modules(root_dir)
-        return
-    for root_module in ROOT_MODULES:
-        module_dir = root_dir / root_module
-        if module_dir.exists():
-            yield from _iterate_modules(module_dir)
-
-
-def _iterate_modules(root_dir: Path) -> Iterator[tuple[Path, list[Path]]]:
-    # local import to avoid circular import
-    from .constants import EXCL_FILES
-    from .loaders import LOADER_BY_FOLDER_NAME
-
-    if not root_dir.exists():
-        return
-    for module_dir in root_dir.iterdir():
-        if not module_dir.is_dir():
-            continue
-        sub_directories = [path for path in module_dir.iterdir() if path.is_dir()]
-        is_any_resource_directories = any(dir.name in LOADER_BY_FOLDER_NAME for dir in sub_directories)
-        if sub_directories and is_any_resource_directories:
-            # Module found
-            yield module_dir, [path for path in module_dir.rglob("*") if path.is_file() and path.name not in EXCL_FILES]
-            # Stop searching for modules in subdirectories
-            continue
-        yield from _iterate_modules(module_dir)
-
-
-@overload
-def module_from_path(path: Path, return_resource_folder: Literal[True]) -> tuple[str, str]: ...
-
-
-@overload
-def module_from_path(path: Path, return_resource_folder: Literal[False] = False) -> str: ...
-
-
-def module_from_path(path: Path, return_resource_folder: bool = False) -> str | tuple[str, str]:
-    """Get the module name from a path"""
-    # local import to avoid circular import
-    from .loaders import LOADER_BY_FOLDER_NAME
-
-    if len(path.parts) == 1:
-        raise ValueError("Path is not a module")
-    last_folder = path.parts[1]
-    for part in path.parts[1:]:
-        if part in LOADER_BY_FOLDER_NAME:
-            if return_resource_folder:
-                return last_folder, part
-            return last_folder
-        last_folder = part
-    raise ValueError("Path is not part of a module")
-
-
-def resource_folder_from_path(path: Path) -> str:
-    """Get the resource_folder from a path"""
-    # local import to avoid circular import
-    from .loaders import LOADER_BY_FOLDER_NAME
-
-    for part in path.parts:
-        if part in LOADER_BY_FOLDER_NAME:
-            return part
-    raise ValueError("Path does not contain a resource folder")
-
-
-def find_directory_with_subdirectories(
-    directory_name: str | None, root_directory: Path
-) -> tuple[Path | None, list[str]]:
-    """Search for a directory with a specific name in the root_directory
-    and return the directory and all subdirectories."""
-    if directory_name is None:
-        return None, []
-    search = [root_directory]
-    while search:
-        current = search.pop()
-        for root in current.iterdir():
-            if not root.is_dir():
-                continue
-            if root.name == directory_name:
-                return root, [d.name for d in root.iterdir() if d.is_dir()]
-            search.append(root)
-    return None, []
-
-
-# Spaces are allowed, but we replace them as well
-_ILLEGAL_CHARACTERS = re.compile(r"[<>:\"/\\|?*\s]")
-
-
-def to_directory_compatible(text: str) -> str:
-    """Convert a string to be compatible with directory names on all platforms"""
-    cleaned = _ILLEGAL_CHARACTERS.sub("_", text)
-    # Replace multiple underscores with a single one
-    return re.sub(r"_+", "_", cleaned)
-
-
-def to_diff(a: dict[str, Any], b: dict[str, Any]) -> Iterator[str]:
-    a_str = yaml.safe_dump(a, sort_keys=True)
-    b_str = yaml.safe_dump(b, sort_keys=True)
-
-    return difflib.unified_diff(a_str.splitlines(), b_str.splitlines())
-
-
-def safe_read(file: Path) -> str:
-    """Falls back on explicit using utf-8 if the default .read_text()"""
-    try:
-        return file.read_text()
-    except UnicodeDecodeError:
-        # On Windows, we may have issues as the encoding is not always utf-8
-        try:
-            return file.read_text(encoding="utf-8")
-        except UnicodeDecodeError:
-            raise
-
-
-def safe_write(file: Path, content: str) -> None:
-    """Falls back on explicit using utf-8 if the default .write_text()"""
-    try:
-        file.write_text(content)
-    except UnicodeEncodeError:
-        # On Windows, we may have issues as the encoding is not always utf-8
-        file.write_text(content, encoding="utf-8")
-
-
-def in_dict(keys: Iterable[str], dictionary: dict) -> bool:
-    return all(key in dictionary for key in keys)
-
-
-def get_cicd_environment() -> str:
-    if "CI" in os.environ and os.getenv("GITHUB_ACTIONS"):
-        return "github"
-    if os.getenv("GITLAB_CI"):
-        return "gitlab"
-    if "CI" in os.environ and "BITBUCKET_BUILD_NUMBER" in os.environ:
-        return "bitbucket"
-    if os.getenv("CIRCLECI"):
-        return "circleci"
-    if os.getenv("TRAVIS"):
-        return "travis"
-    if "TF_BUILD" in os.environ:
-        return "azure"
-
-    return "local"
-
-
-def quote_int_value_by_key_in_yaml(content: str, key: str) -> str:
-    """Quote a value in a yaml string"""
-    # This pattern will match the key if it is not already quoted
-    pattern = rf"^(\s*-?\s*)?{key}:\s*(?!.*['\":])([\d_]+)$"
-    replacement = rf'\1{key}: "\2"'
-
-    return re.sub(pattern, replacement, content, flags=re.MULTILINE)
-
-
-def stringify_value_by_key_in_yaml(content: str, key: str) -> str:
-    """Quote a value in a yaml string"""
-    pattern = rf"^{key}:\s*$"
-    replacement = rf"{key}: |"
-    return re.sub(pattern, replacement, content, flags=re.MULTILINE)
-
-
-def humanize_collection(collection: Collection[Any], /, *, sort: bool = True, bind_word: str = "and") -> str:
-    if not collection:
-        return ""
-    elif len(collection) == 1:
-        return str(next(iter(collection)))
-
-    strings = (str(item) for item in collection)
-    if sort:
-        sequence = sorted(strings)
-    else:
-        sequence = list(strings)
-
-    return f"{', '.join(sequence[:-1])} {bind_word} {sequence[-1]}"
-
-
-class GraphQLParser:
-    _token_pattern = re.compile(r"[\w\n]+|[^\w\s]", flags=re.DOTALL)
-    _multi_newline = re.compile(r"\n+")
-
-    def __init__(self, raw: str, data_model_id: DataModelId) -> None:
-        # Ensure consistent line endings
-        self.raw = raw.replace("\r\n", "\n").replace("\r", "\n")
-        self.data_model_id = data_model_id
-        self._entities: list[_Entity] | None = None
-
-    def get_views(self, include_version: bool = False) -> set[ViewId]:
-        views: set[ViewId] = set()
-        for entity in self._get_entities():
-            if entity.is_imported:
-                continue
-            view_directive: _ViewDirective | None = None
-            for directive in entity.directives:
-                if isinstance(directive, _ViewDirective):
-                    view_directive = directive
-                    break
-            if view_directive:
-                views.add(
-                    ViewId(
-                        view_directive.space or self.data_model_id.space,
-                        view_directive.external_id or entity.identifier,
-                        version=view_directive.version if include_version else None,
-                    )
-                )
-            else:
-                views.add(ViewId(self.data_model_id.space, entity.identifier))
-        return views
-
-    def get_dependencies(self, include_version: bool = False) -> set[ViewId | DataModelId]:
-        dependencies: set[ViewId | DataModelId] = set()
-        for entity in self._get_entities():
-            view_directive: _ViewDirective | None = None
-            is_dependency = False
-            for directive in entity.directives:
-                if isinstance(directive, _Import):
-                    if directive.data_model:
-                        dependencies.add(directive.data_model)
-                        break
-                    is_dependency = True
-                elif isinstance(directive, _ViewDirective):
-                    view_directive = directive
-            if is_dependency and view_directive:
-                dependencies.add(
-                    ViewId(
-                        view_directive.space or self.data_model_id.space,
-                        view_directive.external_id or entity.identifier,
-                        version=view_directive.version if include_version else None,
-                    )
-                )
-            elif is_dependency:
-                # Todo: Warning Likely invalid directive
-                ...
-        return dependencies
-
-    def _get_entities(self) -> list[_Entity]:
-        if self._entities is None:
-            self._entities = self._parse()
-        return self._entities
-
-    def _parse(self) -> list[_Entity]:
-        entities: list[_Entity] = []
-        entity: _Entity | None = None
-        last_class: Literal["type", "interface"] | None = None
-
-        parentheses: list[str] = []
-        directive_tokens: _DirectiveTokens | None = None
-        is_directive_start = False
-        is_comment = False
-        tokens = self._token_pattern.findall(self.raw)
-        for no, token in enumerate(tokens):
-            if no >= 2 and (tokens[no - 2 : no + 1] == ['"'] * 3 or tokens[no - 2 : no + 1] == ["'"] * 3):
-                is_comment = not is_comment
-            if is_comment:
-                continue
-
-            token = self._multi_newline.sub("\n", token)
-            if token != "\n":
-                token = token.removesuffix("\n").removeprefix("\n")
-
-            if token in "({[<":
-                parentheses.append(token)
-            elif token in ")}]>":
-                parentheses.pop()
-
-            is_end_of_entity = bool(parentheses and parentheses[0] == "{")
-            if entity and is_end_of_entity:
-                # End of entity definition
-                if directive_tokens and (directive := directive_tokens.create()):
-                    entity.directives.append(directive)
-                directive_tokens = None
-                entities.append(entity)
-                entity = None
-            elif entity is not None:
-                if directive_tokens is not None and (not parentheses or token == "@"):
-                    # End of directive
-                    if directive := directive_tokens.create():
-                        entity.directives.append(directive)
-                    directive_tokens = None
-                    if token == "@":
-                        is_directive_start = True
-                elif directive_tokens:
-                    if token == "\n" and "{" not in parentheses:
-                        # Throw away.
-                        continue
-                    # Gather the content of the directive
-                    directive_tokens.append(token)
-                elif token == "@":
-                    is_directive_start = True
-                elif is_directive_start and token in ("import", "view"):
-                    directive_tokens = _DirectiveTokens([token])
-                    is_directive_start = False
-                elif is_directive_start:
-                    # Not a directive we care about
-                    is_directive_start = False
-
-            elif token in ("type", "interface"):
-                # Next token starts a new entity definition
-                last_class = token
-            elif last_class is not None and token != "\n":
-                # Start of a new entity definition
-                entity = _Entity(identifier=token, class_=last_class)
-                last_class = None
-        return entities
-
-
-class _DirectiveTokens(list, MutableSequence[str]):
-    def create(self) -> _Directive | None:
-        return _Directive.load(self)
-
-
-@dataclass
-class _Directive:
-    # This pattern ignores commas inside }
-    SPLIT_ON_COMMA_PATTERN = re.compile(r",(?![^{]*\})")
-
-    @classmethod
-    def load(cls, content: list[str]) -> _Directive | None:
-        key, *content = content
-        raw_string = "".join(content).removeprefix("(").removesuffix(")").replace("\n", ",")
-        data = typing.cast(dict[str, Any], cls._create_args(raw_string))
-        if key == "import":
-            return _Import._load(data)
-        if key == "view":
-            return _ViewDirective._load(data)
-        return None
-
-    @classmethod
-    def _create_args(cls, string: str) -> dict[str, Any] | str:
-        if "," not in string and ":" not in string:
-            return string
-        output: dict[str, Any] = {}
-        if string[0] == "{" and string[-1] == "}":
-            string = string[1:-1]
-        if string[0] == ",":
-            string = string[1:]
-        if string[-1] == ",":
-            string = string[:-1]
-        for pair in cls.SPLIT_ON_COMMA_PATTERN.split(string):
-            stripped = pair.strip()
-            if not stripped or ":" not in stripped:
-                continue
-            key, value = cls._clean(*stripped.split(":", maxsplit=1))
-            output[key] = cls._create_args(value)
-        return output
-
-    @classmethod
-    def _clean(cls, *args: Any) -> Any:
-        return tuple(arg.removeprefix('"').removesuffix('"').removeprefix('"').removesuffix('"') for arg in args)
-
-    @classmethod
-    @abstractmethod
-    def _load(cls, data: dict[str, Any]) -> _Directive: ...
-
-
-@dataclass
-class _ViewDirective(_Directive):
-    space: str | None = None
-    external_id: str | None = None
-    version: str | None = None
-
-    @classmethod
-    def _load(cls, data: dict[str, Any]) -> _ViewDirective:
-        return _ViewDirective(space=data.get("space"), external_id=data.get("externalId"), version=data.get("version"))
-
-
-@dataclass
-class _Import(_Directive):
-    data_model: DataModelId | None = None
-
-    @classmethod
-    def _load(cls, data: dict[str, Any]) -> _Import:
-        if "dataModel" in data:
-            return _Import(data_model=DataModelId.load(data["dataModel"]))
-        return _Import()
-
-
-@dataclass
-class _Entity:
-    identifier: str
-    class_: Literal["type", "interface"]
-    directives: list[_Directive] = field(default_factory=list)
-
-    @property
-    def is_imported(self) -> bool:
-        return any(isinstance(directive, _Import) for directive in self.directives)
