+++ cognite_toolkit/_cdf_tk/utils/cicd.py
@@ -1,0 +1,18 @@
+import os
+
+
+def get_cicd_environment() -> str:
+    if "CI" in os.environ and os.getenv("GITHUB_ACTIONS"):
+        return "github"
+    if os.getenv("GITLAB_CI"):
+        return "gitlab"
+    if "CI" in os.environ and "BITBUCKET_BUILD_NUMBER" in os.environ:
+        return "bitbucket"
+    if os.getenv("CIRCLECI"):
+        return "circleci"
+    if os.getenv("TRAVIS"):
+        return "travis"
+    if "TF_BUILD" in os.environ:
+        return "azure"
+
+    return "local"
